Feat: 싱글모드 api 변경에 따른 로직 변경
- 싱글모드 api 변경에 맞춰서 전체 로직을 수정하였습니다
- 기존의 컴포넌트는 그대로 사용하였지만 로직은 완전히 백에서 게임정보를
  받아와서 렌더링하는 로직으로 변경되었습니다
- 직접 테스트까지 완료했는데 아직 미흡한 부분이 존재함

작업하면서 백엔드에서 추가로 처리해줘야 하는 부분
- 게임 시작, 게임 이어하기 api response에 year 추가
- 라운드 종료 response에 준비된 모든 문제가 소진  되었는지 여부 추가
- 게임종료 api response에서 경험치(exp) 추가
- 게임 이어하기로 시작했을 때 기존의 게임이력에 라운드가 종료되었으면
  다음 라운드로 올리는 로직 추가
- 싱글모드 플레이 로그 api에서 request에서 roomId 제거로직 추가

가 모두 적용되면 추후에 프론트도 맞춰서 로직변경 할 예정

Resolves: #14
